requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
beautifulsoup4==4.11.2 ; python_full_version >= '3.6.0'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
diff-match-patch==20200713 ; python_version >= '2.7'
django==4.1.7
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-filter==22.1
django-import-export==3.0.2
django-js-asset==2.0.0 ; python_version >= '3.6'
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0 ; python_version >= '3.6'
markuppy==1.14
odfpy==1.4.1
openpyxl==3.1.1
psycopg2-binary==2.9.5
pyyaml==6.0
soupsieve==2.4 ; python_version >= '3.7'
sqlparse==0.4.3 ; python_version >= '3.5'
tablib[html,ods,xls,xlsx,yaml]==3.3.0 ; python_version >= '3.7'
xlrd==2.0.1
xlwt==1.3.0
